scanner: Add a configurable file archiver for scanned files
Use a `FileArchiver` in the scanner to archive scanned files. By default
archived files are stored in ~/.ort/scanner/archive but this can be
overridden in the configuration.

By default only license files are archived as they will be used in the
notice file generator in a future commit.

TODO
- Update documentation
- Add tests
- Test how the file storage configuration can be reused in HOCON (at
  least partly)
- Rebase on top of file saver commit

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@here.com>