StaticHtmlReporter: Add exemplary links to where detected license were found
For non-excluded findings in a VCS, link each detected license to the
first finding for its specific license type. Due to UI restrictions it
would be difficult to link to all findings for a specific license type,
but usually it is enough to show that a specific license type has been
found at least once, anyway. For similar reasons do this only for the
static HTML report for now, just to briefly demonstrate the general
feature.

Signed-off-by: Sebastian Schuberth <sebastian.schuberth@bosch-si.com>