Refactor VcsHost from a sealed class to an enum
The original idea was that VcsHost should be able to contain state to
talk to a specific instance of a host, e.g. a corporate GitHub
Enterprise server. However, that approach turned out to be
over-engineered for the current use-cases. To simplify the code, use an
enum class instead, and add a tests.

Signed-off-by: Sebastian Schuberth <sebastian.schuberth@bosch-si.com>